Removed a huge memory leak in reading SAM caused by a bug in the thread-safe modifications.
TO FIX STILL:
Backed out the freeing of data in bam_parse_header due to attempting
to reuse it later. I'm not sure of the cause yet and these data
structures need a good reorganisation. For now we'll accept a minor
leakage of RG fields as a tradeoff for working.