Fixed a memory leak in the new cram_block alternatives to dstring_t.
Corrected the storage of ambiguity bases. When encoding N vs R we were
using a DS value of 4 (outside the spec). We now generate 'B' code
features and store the base/qual verbatim.

This meant reordering things a bit, but I've verified I can decode it
with this C code and with the Java code.