Added a -X option to use embedded references.
When encoding, this stores reference portions per slice.

When decoding, this specifies the lack of a reference on the command
line or header is no longer an error. It also uses the embedded
reference in preference to any it finds via other means. (Note it is
legal to have a file where some portions have embedded references and
some do not.)

Removed SM from the preservation map for 1.0 output. While legal in
the 1.0 spec, the Java cramtools didn't implement it so it causes
compatibility issues.

Also fixed some more hard coded block content IDs, replacing them with
#define symbols.

Finally, added some more rigorous memory allocation error checking.