Purged CTF from the source tree.
It had no explicitly stated licence (although was implicitly assumed
to be given to us under the same licence as the rest of the code,
this is not technically enough).

Obviously we could have got this corrected, but the code is also
superfluous (I never saw a CTF file in the wild) and we remove a large
number of compilation warnings too.