A PROOF OF CONCEPT.
Added the ability to load shared libraries for adding custom codecs to
CRAM files.

At present it'll try any it finds in CRAM_CODEC_DIR when encoding,
with no way to control which it uses. (It'll try them all and pick the
best.) In decoding obviously it uses whatever the file was encoded
with.

This needs more fleshing out. Specifically 1 byte is too short for a
codec ID (I suggest 4), there needs to be a codec registry, and we
need a way of controlling on the command line which codecs should be
tried.