Check for the @HD SO:sort-order field, with new parse function in sam_header.c.
We now use this when setting fd->unsorted so that a packed slice with
multiple references in it is not always considered unsorted.  If the
file claims to be sorted then it'll incr/decr refs internal to the
slice as it progresses.

This avoids an issue with high memory usage on very small files where
all data is packed into one slice.  Previously it remembered all
references used as it assumed they'd be reused again very soon
(unsorted mode).