Fixed the broken "special case" in index querying.
The binary search due to the comparisons may yield a value 1 bin too
low when the position asked for precisely matches the start coordinate
of an index bin. The intention of the special case is to correct this,
but it did so without checking whether the current bin returned
actually did overlap the requested range too.

This therefore lead to starting to decode too late into the file,
giving fewer overlapping sequences than desired.