Better tests for bit-based codecs running out of input.
"american fuzzy lop" found a case where the checks for running out of
input added in #3895 failed.  This happened due to the result of
in->uncomp_size - in->byte being unsigned rather than signed, so the check
failed when in->byte was greater than in->uncomp_size.

The checks are pulled into a new static inline cram_not_enough_bits.  This
has been checked with frama_c to work as long as:
   0 <= blk->bit < 8
   0 <= blk->uncomp_size
   blk->uncomp_size, blk->bit and nbits are of type int32_t
   blk->byte is of type size_t

Instances of the old code are replaced with cram_not_enough_bits.

get_one_bits_MSB gets the same checks for running out of input as are already
in get_zero_bits_MSB.

cram_subexp_decode_init gets a check to ensure subexp.k >= 0 and
cram_subexp_decode checks for i >= 0.