Improved error handling with corrupt files.
This is related to the recent htslib change, however done differently
here.

The calling API for scram_get_seq returns -1 for failure to read, but
the caller then has to check cram_eof to determine if that is due to
EOF or not.  The problem is if we deliberately only decode some and
not all then we're not at EOF.

Rather than specifically require the caller to track all this, we now
make cram_eof return -1 if an error has been detected. (In htslib this
is done by making sam_read1 return -2 instead.  Possibly we could do
the same here as well.)