Remove unnecessary comparisons added in revision 3664.
Revision 3664 fixed a problem where some external blocks were not being
uncompressed when they should have been.  Part of this change involved 
comparing the block cram_content_type with values from cram_encoding, which
is not valid.  The real solution to the problem was elsewhere in revision
3664, so this patch removes the incorrect comparisons.  They were harmless
but caused a lot of compiler warnings.