Fixed bug with non-reference base CRAMs and indexing.
When using --output-fmt cram,no_ref the reference sequence is not
loaded into memory, causing the "ref_end" container field to be unset.
This in turn meant that the code that clipped the container header
start/end fields (pos+span) to the actual reference length, to work
around buggy aligners that positiong data off the end of the
reference, was itself buggy.