Minor tweaks.
Use both Z_DEFAULT_STRATEGY and Z_FILTERED only on compression level 6
and above as it's 6-7% more cpu and rarely makes much difference
except on specific data sets.

Also noticed and fixed a cram version 2.1 vs 3.0 difference that
wasn't correctly honoured - the container record counter is 64-bit, so
we need to allow for this.  (In practice it wouldn't make a difference
until we got to > 2 billion records.)