Fixed bugs causing scram_flagstat -t2 to crash.
On error, we shouldn't do t_pool_flush and cram_flush_result as the
data structures may not have been initialised in all threads
(particularly one generating an error).

The second issue is the error itself.  This was due to additional
error checking in r3916 to check unmapped reads have no alignment
position.  It's not a valid check when we are not decode CRAM_AP data
series.