Disable the per-slice md5 creation and checks (SD and BD slice tags).
Behaviour of these have changed in a recent Java Cramtools release,
indicating that we shouldn't be using these until they are formally
added to the specification as they are too fluid currently.

For future work, once CRAM version 3.1 exists with the tags defined
we'll need to only check these tags if the CRAM version is 3.1 and
above as we have generated several v3.0 CRAM files with differing
BD/SD endianness.