Horrid temporary hack to avoid exporting append_int/append_uint if it looks like we're including this file from within gap5.
It's simply a way to avoid a type clash with the same function name in
gap5 and io_lib, but with different signedness prototypes.  Previously
the current tarball of Staden Package doesn't compile cleanly against
the newer tarballs of io_lib.  The ideal solution is to release a
fixed Staden Package too that uses the newly added io_lib function,
but that comes later.