Added a lossy read-name method to scramble (-n).
This takes care to handle supplementary reads too and also uses the TC
aux field if present (it almost never is!).

Note, using scramble to convert a lossy read-name CRAM to a new CRAM
will increase the size as it doesn't currently remember the lossy
mode.  This is either a feature or a bug, depending on use.