cram_write_block() no longer computes the crc32 if it already has one. This is to permit read_block/write_block loops to be efficient.
HOWEVER: this is a user visible change.  Code which does read_block,
edit_block, write_block will now have to zero out the block crc32
value so it is recomputed.  This was the case in cram_write_SAM_hdr.

Also made the fd->ignore_md5 variable be reused for ignoring crc too,
as it's a general ignore checksum option.  Perhaps we need to use
fd->ignore_chksum instead or have a fd->ignore_crc option?
(Probably best to just merge everything into a single option and be
done with it - check vs no_check.)