Reverted the last os.h.in change and applied a better fix.
I realised that HAVE_CONFIG_H may not be set by external tools linking
against io_lib.  Instead I made up a NO_AUTOCONF macro so that it can
be explicitly set for users that cannot use autoconf (and manually
move os.h.in to os.h).

This makes it easier for an automated build system to compile with
outside of autoconf, but without affecting the remainder of us.