Added use of mallopt M_TOP_PAD parameter and appropriate autoconf checks.  Initialisation of this is via the scram_init() function which may be called from client programs. (Or alternatively set the MALLOC_TOP_PAD_ environment variable to ~10 million.)
This avoids an inefficiency of CRAM decoding under GNU malloc.  Cram
decoding dynamically allocates memory for each block in CRAM file,
decodes it, and then frees the blocks.  If these malloc/frees happen,
by happen-stance, to end up at the end of the heap then GNU malloc
will call brk/sbrk to reduce the heap size.  The next CRAM slice then
has to brk again to increase it again.  This leads to a high degree of
page faults.

By setting M_TOP_PAD to a rather arbitrary 10Mb we inform malloc that
it shouldn't give back anything smaller than 10Mb to the system,
instead keeping it incase it needs it in a subsequent malloc.  The
performance gain in decoding CRAM in this situation can be as much as
15%.

Is it worth adding a caching layer to keep old blocks ourselves?
Maybe, but they're all variable in size so it is complex to manage
this efficiently and it gets even more complex once we add
multi-threading into the mix.  Besides this is essentially what
malloc/free are doing anyway.