Added CRC checking for BAM files.
Note this is in the BAM specification, but almost no implementations
have checked it to date!  This includes Scramble, Samtools/Htslib,
Sambamba, CanSam, Picard, Biobambam and cl-sam.  The only exception is
bamsormadup in Biobambam (using a later BAM reader) and tools that
simply pipe BAM into gzopen() wrappers such as bwa aln.

The speed impact is dependent on the task.

Reading (and discarding contents of) a compressed BAM is around 9%
slower.  Converting from compressed BAM to compressed BAM is barely
noticable as compression is the slowest component.  Pure reading 
of uncompressed BAM is around 120% slower, but if we then output
uncompressed BAM again it's only around 18% slower as the encoder has
more work to do and also needed to checksum anyway.

Hence it may be useful to disable validation in some circumstances.
The BAM_OPT_IGNORE_CHECKSUM (or CRAM_OPT_IGNORE_CHECKSUM when calling
scram_set_option) will disable the CRC checking.  This is enabled with
the -! option to scramble and scram_flagstat and previously observed
only by the CRAM decoder.

Also changed the CRC function for the encoder side (outputting BAM) to
use iolib's copy of Stephan Brumme's & Bulat Ziganshin's CRC
implementation.   The combined effect of this is that BAM(uncomp) to
BAM(uncomp) transcoding is still actually faster now, with CRC
checking, than it was before.  So in combination this change is
probably a speed gain.