The SAM/CRAM header blocks are now kepted in the type order they first occur in.
Note that this refers to the first occurence of each type, so a block
of @RG followed by a block of @SQ will appear in that order (and vice
versa if the blocks are swapped), but interleaving @CO between every
line would still cause the @CO records to be grouped together.

Patch supplied by Joe Rayner (Illumina), with slight ammendments by
myself.