Fixed a compression inefficiency when switching to unsorted mode.
We switch from sorted to unsorted mode only after a couple tiny
containers have been created. (Ideally we'd detect upfront.)

We also compute compression metrics on the first few containers and
then keep those stats for the next 100 or so.  The combination of
these meant we computed compression metrics based on data that was not
of comparable size to the rest of the container.  In one test set this
meant Z_RLE was optimal on the 1-read slices but then applied to
10,000 read slices when Z_FILTERED is preferable (due to lots of
duplicate entries).