Fixed a rare renormalisation bug in the rANS codec.
The symbol frequencies need to sum to TOTFREQ (4096 currently) and are
rounded up/down accordingly.  The combination of integer rounding
means the renormalised frequences don't always total 4096 exactly, so
the remainder is added-to / subtracted-from the most frequent symbol.
In one particular data set this remainder was larger than the most
frequent symbol, causing it to become negative.

We now just do another round of renormalisation with slightly lower
products until we get it right.  It's not the fastest solution, but a
very rare event.