Remove rogue file.fa.gzi error.
The previous commit to the bgzip index loading code added a perror,
but this code is called in a "try it and see" mode so the error
message is produced even for cases where we expected it.  This change
isn't ideal, but it no longer complains for the working case.