Improve dstring_resize memory usage.
If we try to resize by a large amount then we just use that size
verbatim without upscaling to the next power of 2.  When resizing by
small incrememnts the power of 2 boost still applies.

This reduces memory usage when (for example) using dstring_append on
an entire buffer / file in one step.