Multi-thread CRAM fixes.
1. Decoding: qwhen using more than 1 slice per container, we need to
move various data items out of the container struct and into the slice
struct if we wish to permit more than one slice (in the same
container) to be decoded in parallel.

See https://github.com/samtools/htslib/pull/666 for additional
information.

2. Decoding: fd->no_ref is set while decoding the compression header
and used in subsequent slice decodes, but we may then decode the
next container compression header before the previous slices have
finished decoding.

3. Decoding: avoid race when limiting data via required_fields.

fd->decode_md is now copied to s->decode_md, so the slice can disable
this itself if required (such as when the user asked for MD tag to be
filled out while also asking not to return any auxiliary blocks).

Although technically fixing a threading violation, the practical
implementations means this is just a tidyup rather than any real
behavior changes.

4. Encoding: Cram_encode_aux needed an extra guard surrounding the
fd->tags_used field.  This is used to hold tag types seen so far in
the file, within any container or block, so we can keep track of the
compression methods that work best for a any specific tag type.