Added an alternative variable length coding mechanism.
This uses 7-bits at a time with top bit to indicate more bytes
needed.  It's totally incompatible with itf8/ltf8 but has other
advantages (one mechanism instead of two, simpler, 0.1% space
saving).  However it's disabled for now as doing it properly for V4
only requires 100s of extra version checks to be added.