Added more FQZ default params.
These work better on MiSeq and IonTorrent data sets.
For now these only get explored with scramble -7.  We still need to
create some fast way of automatically picking the most likely set of
params (for fast work) and can keep the existing way of checking all
for the slower -7 and above.