Make the random number generate for ce#sorted.sam fixed.
I am unsure right now why occasionally the tests fail, but it is
always this file and it is created by generate_data.pl.  I cannot see
how the same instance could ever generate a different data set though,
so this may be a red herring.

On the other hand, it may be parallel vs serial test harness, which
was fixed elsewhere in this branch.