Switch to using adaptive arith instead of rANS4x16.
We may still want rANS4x16 for simple modes (it's faster due to 16-bit
renormalisation), but as an experiment if we use it adaptive modes
everywhere then we can get smaller files.

Generally it's around 1% less, but varies by data.

TODO: consider which ones are worth keeping.  Eg fast rans O0/O1 and
arith with RLE?