Updated fqzcomp to match new V4 spec draft.
There are a lot of changes in this, but primarily it is to permit
multiple parameter blocks instead of a single one, with a selector
model to choose.  (This takes over from the strand model.)

The TEST_MAIN command line interface has had a huge overhaul too.  It
now wants line oriented ASCII quality data with an optional selector
ID.  This makes experimentation easier outside of Scramble proper.