Adjusted fqz parameter locations.
HAVE_QMAP shifted to adjacent to the HAVE_?TAB flags.
DO_REV has now become a global flag instead of per-param block as it
is mainly a CRAM3 vs CRAM4 flag and not data specific.

Also fixed the qmap writing so it doesn't have the duplicate
nsym(max_sym) field.  This means the code is now interchangeable with
the Javascript one, and more importantly matches the spec.