Replaced X_DICT with X_EXT.
The DICT method is complex and only worked on a few specific data sets.

Instead we use the bit flag to indicate an "external" compression tool
instead of the arithmetic coders.  This can be useful from the name
tokeniser (it helps novaseq names over DICT) and permits use of
arith_dynamic for transforms, eg X4 and PACK with bzip2.

Possibly this should be done in a more general purpose way by
separating transforms from compressors, but that's moving beyond
simple CRAM 3.1 and into 4.0 space.