Bug fix auto-tuning of rANS methods.
The compression metrics culls methods that aren't useful over time,
however if we selected RANS_PR193 as the only suitable method and then
subsequently spot we have more than 16 symbols (where bit-packing is
disallowed) then we don't have a fallback method.  This lead to an
assertion failure.

(We already coped with this in all other places bar one, similarly for
ARITH_PR193.)

Also fixed cram_dump to show the original (raw) compression method in
addition to the internal variant of that method.