CRAM4: dedup read names.
We use the same mate_line data to avoid duplicating read names.  This
is dedup via future copy which isn't ideal.  So the first read that
claims to be not-detached with mate-downstream has no read name and
copies from the (downstream) mate.  Perhaps not the most efficient
orientation, but the downstream mate doesn't have the mate-line field
present.  My code happens to know this, but it may not be easy for all
implementations to work in that order.