CRAM 4: new EXPLICIT_TLEN CF flag.
This is set when a pair is in the same slice and mate flags, rnext and
pnext are consistent between all template copies, but tlen is not.  Eg
the old out-by-1 problem or the still present 5'/3' vs left/right-most.

Generally it's a tiny gain, but on some older data sets it can save
~5% of storage.