CRAM 4: Store MD*, NM* and RG* in TD map.
These store the positions of MD, NM and RG tags but do not store the
actual value verbatim (assuming it's not needed due to MD/NM being
wrong).  Thus the output stream now knows whether or not the field
should be auto-generated and it also puts it in the correct place,
permitting bit-identical SAM->CRAM->SAM round trips with these
fields.