Experimental TLEN0 flag (commented out).
This is a flag that indicates RNEXT/PNEXT/TLEN of */0/0 and still
permits pairing of reads for flag and name dedup purposes.  It's not
finished though as it isn't used by the decoder and I'm not sure it
works properly when one end of as pair is */0/0 and the other is not.