Rewrote EOF writing code.
This creates the blocks as intended rather than a raw byte stream.
It's easier to understand and avoids manual effort to figure out
the CRC values.

The only reason we had explicit byte streams written was due to lack
of clarity in the ITF8 specification at the time about what the bottom
4 bits should be on a 32-bit (4.5 byte) number.  C had all zeros while
Java had all 1.

The C code generates the correct byte stream given its native
encoding, so this code is sufficient.