Added signed LONG/INT types.
AP (alignment pos) and TS (template size) both use signed long.
RG (read group) has signed int as it can be -1.

For BETA we must require the values to be >= 0, so use offset
accordingly.

For HUFFMAN we amend the cram_huffman_encode_store function to cope
with signed vals.  The bit encoding from them on doesn't change.

The main work is for EXTERNAL