Fixed more signed vs unsigned issues with CRAM4.
Looking for 0x8f 0xff 0xff 0xff 0x7f in an uncompressed CRAM binary
shows where we are attempting to write -1 as an unsigned 32-bit
quantity.  We don't want this possibility as it prevents us from
changing the type to a 64-bit quantity later.  We want all such
quantities to be marked as signed in the spec and use zigzag
encoding to be data-size agnostic.

- Ref_seq_id in container and slice headers are now signed.

- Embebbed ref id is now 0 for unused instead of -1.  (0 is CORE block
  so was never a valid external ID anyway.)

- RI (ref ID) and NS (mate ref ID) may be -1, so are both now signed.

INCOMPATIBILITY:  This is incompatible with previous versions of
scramble if using the experimental CRAM4.0 format.