CRAM4 addition of EXTERNAL_SIGNED and HUFFMAN_SIGNED.
These migrate the sign of a field from the specification tables to the
encoding specified, making both the decode and spec cleaner to
describe.

Removed use of E_BETA for unsorted positions in CRAM4 so we can
cull all usage of CORE block.

Also fixed a minor memory leak in cram_load_reference when given an
invalid reference file name.