Added an io_lib/version.h header file.
The header has macros for the string version (IOLIB_VERSION) along
with major, minor and patch version numerics. This replaces the need
to use io_lib_config.h which should never have been installed.

In order to validate this, I tested a build of progs directory using

    make DEFAULT_INCLUDES="-I/tmp/inst/include" \
         CPPFLAGS="-UHAVE_CONFIG_H" AM_CPPFLAGS=""

Although not strictly necessary as progs is part of the core package,
I wish it to be a demonstration of how to use the library without
requiring internal knowledge of the sibling io_lib directory.

This required a few extra HAVE_CONFIG_H guards along with the obvious
renaming of PACKAGE_VERSION to IOLIB_VERSION in several places.

Fixes #25