Improve CRAM writing speeds.
The bam_copy function copied bf->alloc bytes rather than the actual
used number of bytes.  Given a normal reuse of bam objects, the alloc
size is often the maximum size of any previously observed BAM struct
rather than the used size of this particular struct.

On a test set with some wildly varying lengths of XA:Z tags, this
meant it was copying around 8K each time instead of < 1K.  It makes
little difference to a single thread, but this is in the main thread
and scramble -t16 became bottlenecked in the main thread.  This fix
more than doubled throughput in that specific example.

The +44 is a bit hairy, but it comes from the bam.c code.  It's the
additional size of bam_seq_t structure elements that aren't accounted
for in the BAM blk_size field.