Improved BAM's long-cigar functionality.
This was originally implemented in io_lib using the bin field to
extend the CIGAR from 16-bits to 32-bits and an extra FLAG.

A couple years later the adoption of long-cigars in the SAM spec chose
a different route of using a CG auxiliary tag.

We now write out BAMs using the official CG tag method, but still read
both methods and internally in memory it is still using the former
(with bin be recomputed on-the-fly when writing, so as before don't
trust the memory struct).