Fixed CRAM decoding of feature code 'q'.
Also fixed initial setting of quality values when using lossy quality
mode.  In this mode if nothing edits qualities, they start off at qual
255 (as in BAM), which produces qual string "*".  If we using Q, B or
q features then at least some of the qualities are now set.  We were
not resetting the other quals from 255, but we are now emitting a
quality string.  This gave erroneous quality values for the remaining
bases.  We use Q30 ("?") as the new default in that scenario, which is
the same as htsjdk.

See https://github.com/samtools/htslib/pull/1094 for more details.