Added libdeflate to CRAM.
This does help somewhat, although it needs care for read names where
libdeflate does poorly.

Also tweaked level 1 for CRAM 3.1 so it doesn't do the name
tokeniser.  This has a significant speed reduction when using
libdeflate (as it's lightning fast at low levels), but also costs some
in ratio.  That's appropriate for level 1 which should always favour
speed.

Tested on NovaSeq data, the difference between this PR and master for
CRAM v3.0 and CRAM v3.1 (with libdeflate enabled) is:

Scramble V3.0
     this PR               master
1    27689  232299443      28413  243735921   -/---
2    28186  232070075      28612  240836639   -/--
3    28111  229943268      29539  236890766   -/--
4    29825  226534290      31851  226728661   ./-
5    33884  223761647      34817  223668364   -/.
6    37302  222835285      38071  223030805   -/.
7    51380  220518134      47044  221623486   +/-
8   131938  217898510      67969  220398723   ++/--
9   173147  216836918      77280  220395302   +++/--

Scramble V3.1
     this PR               master
1    25232  219257365      33166  211611200   --/++
2    31337  198332765      30911  198903857   -/-
3    33196  186907583      33070  187474053   ./-
4    33557  186519865      33991  187084328   ./-
5    34039  186429583      35656  186188618   -/+
6    34430  186331049      34742  186314028   ./.
7    43951  185700327      40936  186196808   +/-
8    56630  185373641      48476  186047278   ++/--
9    65647  185299407      56469  186046055   ++/--

High compression levels with 3.0 cost dearly as libdeflate is
expensive at top levels, but gives a good ratio.  Like -1 being
very speed focused, it's fair for -9 to be very size focused and
inappropriate for -9 to still be considering speed overly much.

Internally we use a mix of default (-5) and -7 if we wish to aim for a
slighter slower sweet spot.