Added int64 versions of some HashTable functions.
This is needed for CRAM 4.0 which uses 64-bit integer keys in the hash
table, which doesn't fit in the char*key parameter on 32-bit systems.
This does however add a small dependency on a single C11 feature
(anonymous unions), but we'll see if this causes problems.  Certainly
fewer problems than changing the API, so it's a starting point.

Fixes #31