Fix MacOS tests.
The sed script to sanitise @PG lines causes an failure on MacOS when
running on ce#large_seq.sam.  Likely it's a line length problem
(sigh!).

It's been replaced with a simple grep -v.  We don't need to be so
overly cautious is checking PG line validity so just nuking them all
is simpler.