Fix a CRAM reference threading bug when fetching ref from fasta file.
This copies over a fix from samtools/htslib@47da3929 which tracks
references differently depending on whether they have come from a
fasta file or an md5sum location.  (If we clear the .fa reference then
when it attempts to load it again we've lost the filename and instead
attempt (and fail) to fetch from the @SQ UR filename header.)

Fixes #46.