Add use of FQZComp_qual SEQ context for PB/ONT data.
This is a bit messy because the cram_decode_seq code automatically
decodes quality at the same time.  If sequence is used as a context
for quality decoding, then that code needs separating.

Our solution is to twiddle the required-fields parameter to disable QS
decoding, and then explicitly decode it afterwards.  Note this only
works if we're not using any of the compound seq/qual CRAM feature
codes, but I think we'd likely have to forbid their usage anyway
within the specification as we'd need len(seq) == len(qual) for the
context to make any sense.

An example of a local PacBio file with 100k records:

Before: 187305340 bytes (145419949 QS)
After:  165605250 bytes (123719859 QS)