Add some commented out code for fixing seq_span in unmapped data.
samtools/htslib#1387 identifies a bug where htslib's index query fails
to return data if span=0 (as start+span-1 is -1).  The spec states
than span must be 0 in this case, but no one has yet complained about
writing out 1 and fixing this will make it more likely to trigger the
htslib bug.

So the fix is noted, but not enabled (for now).