Short term solution for missing declaration of encode_names in libhtscodecs2 version 1.3.0
Libhtscodecs2 version 1.3.0 renamed e.g. function encode_names to
tok3_encode_names.  It added a compatibility function encode_names,
but does not declare it in the header file.  Therefore in the build of
staden-io-lib the function encode_names is implicitly declared -
therefore seems to default to a return type of int - and therefore the
returned pointer gets truncated to the lower 4 bytes. This get through
unnoticed until the pointer is freed and produces a segfault there.

Probably the right thing would be to have the short function name in
the header file of htscodecs, therefore solve this issue for other
packages using htscodecs.

This patch implements a short term solution which makes the build and
tests succeed without "implicit declarations".

Fixes Debian bug https://bugs.debian.org/1022282