Silence many buffer overrun warnings in gcc12.
The questionable use of bam_seq_t and &b->data as a placeholder for an
arbitrary number of additional bytes seemed like a reasonable approach
at the time, but was the wrong strategy.  The compilers are now
spotting that the 1 byte "data" field isn't big enough for some
operations.  In theory this may cause optimisation woes too?

Instead we disguise things by using offsetof from the start of the
structure.