Fix CRAM QS values where ambiguity codes are in the reference.
Commit 0aa2af3 changes QS encoding from verbatim copying out of
s->qual_blk and into calling codecs[DS_QS]->encode.  This was part of
the plan to be able to perform the data transforms on QS, included in
CRAM 4.0 prototype.

Unfortunately the ordering was wrong, with a copy of all quals
followed by encoding any explicit QS items produced by the feature
codes.  The correct ordering is the reversal of this.